Fix(에러메세지): 고침 에러메세지
잘못된 내용을 출력하는 에러메세지 문구 변경

+++ src/main/java/com/project/shoppingmall/controller/purchase_retrieve/PurchaseRetrieveController.java
@@ -40,7 +40,7 @@ public class PurchaseRetrieveController {
   public OutputRetrievePurchaseBySeller retrievePurchasesBySeller(
       @PositiveOrZero @RequestParam("sliceNumber") int sliceNumber,
       @Positive @RequestParam("sliceSize") int sliceSize,
-      @RequestParam("productId") int productId) {
+      @RequestParam("productId") long productId) {
     AuthUserDetail userDetail =
         (AuthUserDetail) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
     Slice<PurchaseItem> sliceResult =
